+++ app/proxy/srcs/nginx.conf
@@ -30,21 +30,21 @@ http {
         proxy_pass              http://backend:3000/;
       }
 
-      location /grafana/ {
-        proxy_pass              http://grafana:8000/;
-      }
-      location /phpmyadmin/ {
-        proxy_pass http://phpmyadmin/;
-      }
+      # location /grafana/ {
+      #   proxy_pass              http://grafana:8000/;
+      # }
+      # location /phpmyadmin/ {
+      #   proxy_pass http://phpmyadmin/;
+      # }
 
       # Proxy Grafana Live WebSocket connections.
-      location /grafana/api/live/ {
-        proxy_pass http://grafana:8000/grafana/api/live/;
-        proxy_http_version 1.1;
-        proxy_set_header Upgrade $http_upgrade;
-        proxy_set_header Connection "upgrade";
-        proxy_set_header Host $http_host;
-      }
+      # location /grafana/api/live/ {
+      #   proxy_pass http://grafana:8000/grafana/api/live/;
+      #   proxy_http_version 1.1;
+      #   proxy_set_header Upgrade $http_upgrade;
+      #   proxy_set_header Connection "upgrade";
+      #   proxy_set_header Host $http_host;
+      # }
 
       location /_next/ {
         proxy_pass http://frontend:4000/_next/;
